model/requirements.txt
spotipy
pyspark
nltk
numpy
lyricsgenius
python-dotenv
pandas
beautifulsoup4
seaborn
